Add files via upload
n this Python code, I have implemented a TCP Client-Server application using the socket and thread libraries. The purpose of this application is to create a group chat environment.

The application allows users to create their own group chats. Additionally, new clients can join an existing group chat if they know the correct name and ID of the group. Clients also have the ability to disconnect from the server and leave the group chat if desired.

To use the application, first, you need to run the server. Once the server is up and running, you can launch as many clients as needed to join the group chat.

Enjoy!